+++ CHANGELOG.md
@@ -81,8 +81,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `Disable-SqlDscAudit`
   - Support for debugging of integration tests in AppVeyor.
     - Only run for pull requests
-  - Add new resource SqlServerAudit.
-  - Add new resource SqlServerAuditSpecification.
+  - Add new resource SqlAudit.
+  - Add new resource SqlAuditSpecification.
 - CommonTestHelper
   - `Import-SqlModuleStub`
     - Added the optional parameter **PasThru** that, if used, will return the